git-svn-id: http://svn.geogr.uni-jena.de/svn/jamsmodels/devel@759 503a54e1-1403-0410-a754-803ee147b016

+++ J2K_S_N/src/org/jams/j2k/s_n/J2KNRoutinglayer.java
@@ -260,7 +260,7 @@ import org.unijena.jams.model.*;
             toReach.setDouble("N_RG2_in", NRG2in);
             
         } else{
-            System.out.println("Current entity ID: " + entity.getInt("ID") + " has no receiver.");
+            System.out.println("Current entity ID: " + entity.getDouble("ID") + " has no receiver.");
         }
         
     }
